allow 1pct-20pct coldstart classes to be interdex reshuffable
Summary:
as title, the thinking behind here is that those classes are not as perf sensitive as other classes, so moving them around to save apk size might be a good idea.
Notice that we only plan to try it on IG for now, as we used to run totally disabling interdex class ordering for IG, and still see no perf regression. So this is even more hopeful.

Reviewed By: NTillmann

Differential Revision: D52660673

fbshipit-source-id: ccfba09f38f65933518717e001193cf65489b692