allow interdex reshuffle coldclass classes between 20pct-1pct
Summary:
as title, the thinking behind here is that those classes are not as perf sensitive as other classes, so moving them around to save apk size might be a good idea.
Notice that we only plan to try it on IG for now, as we used to run totally disabling interdex class ordering for IG, and still see no perf regression. So this is even more hopeful.

Reviewed By: NTillmann

Differential Revision: D52607699

fbshipit-source-id: 4f5ec877078377f1c84e06e75103f216005b1fc8