Revert D52663477: Multisect successfully blamed "D52663477: [redex] Promote redex-staging to redex-stable, 2nd attempt" for obuild failure
Summary:
This diff is reverting D52663477
D52663477: [redex] Promote redex-staging to redex-stable, 2nd attempt by NTillmann has been identified to be causing the following build failure:

Tests affected:
- [automation_milan_data_target30_rdk](https://www.internalfb.com/intern/test/562949996670800/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3924924
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Differential Revision: D52720018

fbshipit-source-id: 0cd63c238b0d89ea3ac264a063dd1d790e887daa