Handle mostly empty blocks before calling SwitchEquivFinder
Summary:
CFG simplification works differently in instrumentation mode. As a result, a switch-equivalent method might have an empty-ish block (that is, a block with nothing but a source block and goes to another block). Example: https://www.internalfb.com/phabricator/paste/view/P1009232177?lines=33-35%2C70-73

This situation is unique to instrumentation mode, and the corresponding test case sets up a cfg that represents this.

The reason why this is unsupported by `SwitchEquivFinder` is due to how it tracks "extra loads" ([see comments here for context](https://www.internalfb.com/code/fbsource/[a84cb065124fc7b836af0f7698906d8ad0d754c0]/fbandroid/native/redex/service/switch-partitioning/SwitchEquivFinder.h?lines=169-174)). It represents the state of const values when the dest blocks are reached, and if one block jumps to another without it being parsed/understood, what it represents for "extra loads" could be inaccurate.

This change expands upon the strategy that `SwitchEquivFinder` employs already, which is to within a threshold allow for instructions to get duplicated to avoid misrepresenting things. This is the [transcript of the modification in the test case](https://www.internalfb.com/phabricator/paste/view/P1016775534?lines=47-50%2C60-61), which allows the finder to accurately represent the constant that flows into `INVOKE_STATIC v1, LFoo;.useReg:(I)V` in all cases.

Reviewed By: agampe

Differential Revision: D52740996

fbshipit-source-id: f209f14956c31d26023971834d92767ddbca22b6