make clinit_may_have_side_effects more conservative
Summary:
Consider this:
```
class A {
  static { B.data = 1; }
  int data;
}
class B extends A {
  static { B.data = A.data; }
  int data;
}
```

Even though the `clinit` of `B` looks rather harmless, as it only depends on classes that are guaranteed to have been initialized "earlier" (its super class `A`), it is not universally "side effect free", as the result of its execution depends on whether the static initializer of `A` ran earlier by itself!

Consider the trace for
```
int foo() {
  int x = A.data;
  return B.data;
}
```
```
// (static initialization trace)
  A.<clinit>()
    B.<cinit>()
      B.data = /* A.data */ 0;
    B.data = 1;
```
vs the trace for
```
int foo_wrongly_optimized() {
  return B.data;
}
```
```
// (static initialization trace)
  B.<cinit>()
    A.<clinit>()
      B.data = 1;
    B.data = /* A.data */ 0;
```

This is getting fixed here, in a rather trivial way: Once there has been a side-effecting clinit in a class hierarchy, all clinits of subtypes are considered side-effecting (as we don't currently have the tools to determine whether the super type side effects affects a derived type clinit).

Apparently Google (looking e.g. at the `com.google.android.gms:play-services-fido` library) is fond of this pattern.

Reviewed By: wsanville

Differential Revision: D52878264

fbshipit-source-id: 24de5000bb27805d8bb8874ead5b95d4aac68051