Work around Resolver resolving MethodSearch::Virtual to interface method
Summary:
In `ResolveRefsPass`, we sometimes resolve `invoke-interfaces`s to `invoke-virtuals` to reduce unique ref count. However, we had an issue earlier where we incidentally resolved `invoke-virtual` method ref to an interface method resulting in incorrect bytecode. The issue was addressed with D52631931 at the time.

After digging a bit more, I realized that the issue is in the path where to try to resolve the method ref using type inference information. For an `invoke-virtual` we sometimes resolve the target obj to be an interface type. When passing an interface class to the `Resolver` (with `MethodSearch::Virtual`), it can still resolve it to an matching method on the interface class. We took the result without checking the invoke opcode compatibility. This is hopefully a proper fix that stops the process at an earlier stage.

Reviewed By: ssj933

Differential Revision: D53501687

fbshipit-source-id: a9c29ffb48f4b1b9d0602d0f2ea4ec3ed6affa2b