Change to Node* and links
Summary:
Avoid the need for map lookups when walking the method override graph. This increases the amount of memory needed as well as how the nodes are allocated (and de-allocated).

NOTE: `get_node` returning the empty node means that for methods not in the graph `mog.get_node(method).method != method`!

Reviewed By: beicy

Differential Revision: D53526496

fbshipit-source-id: 6db86ea648441755dc0c9d39b82d630dfaba5f19