make type_class cheap
Summary:
Instead of being backed by an `unordered_map`, which requires locking when being populated, `type_class` is now backed by a field in `DexType`. This should make lookups way faster.

While this adds a field to DexType, we save some space by replacing the old map with a set. So this is largely memory neutral.

No public interfaces changes (except that we add a new convenience method to RedexContext, `class_type`).

Reviewed By: beicy

Differential Revision: D53781470

fbshipit-source-id: eb126a8a757181a19d9c7cda8b4323c3505cb009