Remove abstract/interface assertion (#839)
Summary:
This assertion breaks with trivial examples and appears unnecessary.

### Reviewer note
It appears this assertion can be violated with a fairly mundane setup:

```java
class Internal {
    String process(String input) {
        return "";
    }
}

class External extends Internal {
    Override
    String process(String input) {
        return "";
    }
}

class Issue extends External {
    void foo(String input) {
        new External().process(input);
    }
}
```

where `Internal` and `Issue` are on the classpath, and `External` is on the libpath. We removed this assertion in our fork and would suggest the same change upstream.

Pull Request resolved: https://github.com/facebook/redex/pull/839

Reviewed By: beicy

Differential Revision: D53821890

Pulled By: NTillmann

fbshipit-source-id: 6b18101f4adb7b85605324cd2f80034d9f558272