make deterministic
Summary:
The previous diff introduced some non-determinism: The items of an unordered map are checked in a loop, and a "complete" flag might be set. But other code in the loop also checks if the flag is set.
This is getting separated into two steps here, as it should be, making the checking of the flag independent of the order in which the items are processed.

Reviewed By: beicy

Differential Revision: D54315145

fbshipit-source-id: 833e73b88d44ba9138a542ea79fe28a247d753f9