estimate inlinable methods
Summary:
Even when we decide early to not handle a particular non-escaping new-instance instruction due to obvious size concerns, we can still determine which methods would have been inlinable, and then add those methods to a "methods kept" set. This sharpens the cost function for other inlining opportunities.

This change doesn't do much by itself, but is important for later diffs.

This does not affect the correctness of the transformation, only the cost function.

Reviewed By: beicy

Differential Revision: D52440032

fbshipit-source-id: c49fc21c795c58e2ed05e282cb5a50801c3c1228