allow join with zero
Summary:
It is relatively common that non-escaping objects are joined with zero. Later accesses to such objects are typically guarded in a way to avoid NullPointerExceptions.

Here, we add support for such code. We have to be careful with check-cast and instance-of instructions to reflect the potential null case.

As the `optionalReduceTo42SuppressNPE` test case shows, this code transform may suppress NPEs, as field accesses always succeed, and never throw an NPE. Suppressing NPEs is in line with what other optimizations in Redex do, and what we document. In particular, the RemoveBuilderPatternPass does the same thing.

Reviewed By: thezhangwei

Differential Revision: D53112245

fbshipit-source-id: c9da3664d64e7c5a1bb3181d937dc47bdeb0f83b