support invoke_super across classes
Summary:
We already support invoke-super as part of the analysis in OEA.
There was a remaining limitation in the inliner when the caller and callee are not in the same class. This limitation is lifted here, by temporarily representing such cases in a special way (invoke-direct + virtual method), which eventually must disappear after all inlining steps are done.

Reviewed By: thezhangwei

Differential Revision: D52451716

fbshipit-source-id: 7d76c9a08ca349ae9f282fff8e296a9664070f1f