typedefannochecker: support kotlin annotated properties
Summary:
Support annotated properties in Kotlin. Unlike Java fields, Kotlin properties do not have annotations attached to them. Instead, a synthetic method is created that holds the annotations. To get and set the property value, two more synthetic methods are created too, which have no annotations.

Given a method with `$annotations` in its name, there is probably a Kotlin property and a getter method. It may or may not have a setter method too. Add annotations to the properties and all methods.

Some`$annotations` methods annotate another method instead of a field

Reviewed By: thezhangwei

Differential Revision: D54104892

fbshipit-source-id: 81a6ae8dd0b972809879f609eb039b079fdcf6e8