IRAssembler support for fill-array-data
Summary:
I'd like to start improving some support for array handling like https://fburl.com/code/eabffe38, but first make it possible to write s-expr test cases on the various ways to define an array. Using an integ test is not very nice, as it does not give control on how the array gets populated.

It would be nice to define char[] with actual chars like a b c but I'm not going to bother doing escapes and whatnot. Just simple expression list format with hex values should be good enough.

Reviewed By: agampe

Differential Revision: D53624243

fbshipit-source-id: 9d0881789019da06a86fa0105c42003acbc21472