allow join with zero (fixed)
Summary:
This is an improved version of D53112245:

It is relatively common that non-escaping objects are joined with zero. Later accesses to such objects are typically guarded in a way to avoid NullPointerExceptions.

Here, we add support for such code. We have to be careful with check-cast and instance-of instructions to reflect the potential null case.

As the optionalReduceTo42SuppressNPE test case shows, this code transform may suppress NPEs, as field accesses always succeed, and never throw an NPE. Suppressing NPEs is in line with what other optimizations in Redex do, and what we document. In particular, the RemoveBuilderPatternPass does the same thing.

optionalReduceTo42Override and optionalReduceTo42CheckCast are additional regression tests that show th

Reviewed By: wsanville

Differential Revision: D55051862

fbshipit-source-id: 37d285f67d22b64bcb33eeace1948acd1094589b