Initiate class merging model in InterDexReshufflePass
Summary:
Initiate the needed class merging model in the `InterDexReshufflePass` and turn on mergeability-aware algo in the first run of the pass.
We currently use the merging model only to get the mapping from mergeable classes to their merging types.
In the future, we might use it for more accurate info on which methods can be deduped, and which methods can be merged into the same method, etc.

Reviewed By: thezhangwei

Differential Revision: D55593486

fbshipit-source-id: a26e443300422a86bba54158837ce4dd41e228c9