allow for relaxed constructor and new-instance type assertion
Summary:
I cleaned up the code a bit (according to feedback I gave on the original diff years ago), but more importantly, the assertion is now optionally a little bit weaker, allowing for constructors from supertypes. This is needed for later diffs.
Besides the assertion weakening, this is a behavior-preserving change.

Reviewed By: beicy

Differential Revision: D55705923

fbshipit-source-id: e74f3ee44348160a224178d0416098285ea4a82b