relaxed assumption about init invocations
Summary:
OptimizeEnums assumes that every new-instance of an enum type is followed by an invocation to the enum type's constructor.
We are relaxing this requirement here, by simply excluding all classes where the assumption doesn't hold, as an upcoming diff will allow a new-instance to be followed by a invocation of a super constructor, e.g. the `Enum::.<init>`.

This is a behavior-preserving change if the original assumption holds (which it does so far).

Reviewed By: thezhangwei

Differential Revision: D55775624

fbshipit-source-id: dfa0efa11576ede20afd3da6b98ca8d8f9275322