relaxed assumption about init invocations
Summary:
VerticalMerging attempts to merge a sub-class into a super-class if, besides other conditions, the super-class is never instantiated.
It then assumed that all of the constructor calls to the super-class only occur in the sub-class as part of its constructors.
We are relaxing this requirement here, as an upcoming diff will allow a new-instance to be followed by a invocation of a super constructor.

This is a behavior-preserving change if the original assumption holds (which it does so far).

Reviewed By: beicy

Differential Revision: D55705922

fbshipit-source-id: 941af06ee1a159b1aac3b1f22094bc713f0cc9d9