support for relaxed init inlining
Summary:
The verifier in newer Android versions (min-sdk >= 21) allows calling a super-constructor on newly created objects.
https://android.googlesource.com/platform/art/+/android-5.0.1_r1/runtime/verifier/method_verifier.cc#2274
This got never re-enabled, and R8 also started exploiting this.

This diff introduces a new option that allows the inliner to inline more constructors, producing this newly allowed pattern. It is then left to the inliner's cost function to decide whether it's going to reduce code size overall.
The new functionality is optional, and would only kick in when min-sdk is high enough.
The option is off by default.
So this is a behavior-preserving change.

Reviewed By: beicy

Differential Revision: D55725021

fbshipit-source-id: b6587f2879793d3efea99ef75dbce5c3c40a06f4