Track use of reflected types in mergeability checks
Summary:
# Context

As more and more code has been migrated to Kotlin, more anonymous classes emerged in the order set in the betamap. Since we inject artificial `const-class` instructions in class preload skeleton methods for the big apps in `ClassPreloadGenerator`, we ended up marking those startup set classes as reflected. As a result, we are excluding a larger number of classes as not mergeable.

# The Change

If we actually track the use of the `const-class`es using `DefUseChains`, we will be able to identify the artificial ones and not register them as real reflections.

Reviewed By: ssj933

Differential Revision: D55993834

fbshipit-source-id: ea440ecd17e3f2741503ef0bd3ed4d0fb8221850