Add default 0 value, const folding, and const opcode anno type
Summary:
When a method parameter takes the default value and is an integer, the $default method will receive a CONST 0 instead of the actual default value. In the method, it'll check if the parameter is 0, and if it is, it will call the actual callee with the default value. This is safe, since the checker ensures that the default value is safe.

If we have a case like `val = getBool()? 1 : 0`, javac will optimize that to `val = getBool()`. When dealing with IntDefs, `getBool()` is not safe, but the value could be safe if both 1 and 0 are members of the IntDef.

For cases such as P1223077241, where `v4` is an integer with an annotation `A`, `OPCODE_CONST` will not reset the annotation when we put a different value in `v4`. Fix TypeInference to reset the annotation for `OPCODE_CONST`. This does not show up in `OPCODE_CONST_STRING` since it calls `set_reference` which will set the annotation to `nullptr` https://fburl.com/code/7rqn8hee

Reviewed By: thezhangwei

Differential Revision: D56603421

fbshipit-source-id: 2d669c8fa4f53f96ab87dbd2266aee1451fa7515