Make RenameClassesPassV2::rename_classes_in_layouts safer
Summary: Filter the given rename map before calling `rename_classes_in_layouts()`, so that only class names under the configured packages `allow_layout_rename_packages` are changed. That should be safer and make it less likely for a coincidental rename of a commonly named class like `a` which also happens to be an unrelated string attribute value.

Reviewed By: ssj933

Differential Revision: D56372652

fbshipit-source-id: b7f9e4c6228a40b1b758b1ba8e86eff3fc29549f