Support reading string references from xml layout attributes
Summary: Filter the given rename map before calling `rename_classes_in_layouts()`, so that only class names under the configured packages `allow_layout_rename_packages` are changed. That should be safer and make it less likely for a coincidental rename of a commonly named class like `a` which also happens to be an unrelated string attribute value.

Reviewed By: ssj933

Differential Revision: D56616746

fbshipit-source-id: 7446a770c86ae242ccfd76cb67fc27cf8991dea9