Reorganize functions for checking Java identifiers (#859)
Summary:


I made the misguided choice of refactoring and reusing some methods from ProguardLexer.cpp, and broke the build on Windows for a reason I cannot understand (`TokenType` enum values in ProguardLexer.h somehow were not parsed as valid anymore???? Dunno).

Revert that part of the change, and instead move the useful methods from ProguardLexer to the `java_names` namespace in DexUtil.h.

Differential Revision: D56921682