Reorganize functions for checking Java identifiers (#859)
Summary:
Pull Request resolved: https://github.com/facebook/redex/pull/859

I made the misguided choice of refactoring and reusing some methods from ProguardLexer.cpp, and broke the build on Windows for a reason I cannot understand (`TokenType` enum values in ProguardLexer.h somehow were not parsed as valid anymore???? Dunno).

Revert that part of the change, and instead move the useful methods from ProguardLexer to the `java_names` namespace in DexUtil.h.

Reviewed By: thezhangwei

Differential Revision: D56921682

fbshipit-source-id: bd27c05e66124619c2dff50d418ac447384f7359