Remove debug check for finding classes under other res directories
Summary:
This method scans xml files for strings that could be possible Java class names. Long ago, a perf optimization was put in to limit the directories searched to exclude things that would be unlikely to reference dex code, like color or vector drawables.

In debug mode, there is an assert to run this logic twice, the second time with no directory filters and assert the size match. With the more broad searching done by D56372640, searching the drawable directory reports more strings.

These are not strings that are class names (strings like P1229962922 from P1229963360 are not classes) but nevertheless the assert was firing because at this level of the API, it doesn't check that these strings are actual types.

I think that we can get rid of this assert since D56372640 will put us in a better position to find more real class names anyways.

Reviewed By: jimmycFB

Differential Revision: D56961260

fbshipit-source-id: bc1920645da93f52162c8a4c5295445d6b52fb1a