Option to enable only specific DeepData interactions in Redex
Summary:
We want the ability to add interactions to aggregated DeepData profiles without this necessarily influencing anything Redex is doing. Currently, this is not possible.

This change adds a simple option to `redex.py` to allow filtering the list of interaction profiles passed to Redex for that app (for both method and block profiles).  By default, all interactions are still passed.

Note that this applies to ArtProfileWriterPass, which might not be desirable, so future (more complex) work here may be necessary.

Reviewed By: dcaspi

Differential Revision: D57408430

fbshipit-source-id: 746bfc59efd9009fdbf57a8f6ac1f6157c7e0f3a