Patch fields given constructors with annotated params
Summary:
Some classes are initialized with annotated parameters. The constructor has annotated parameters, but the fields that hold these values are not annotated. If a constructor has annotated parameters, patch the fields so that they're annotated too.

The class might have getter/setter methods that need to be annotated. The kotlin annotations patcher will not patch these methods since there is no `$annotations` method. The annotations come from the constructor. If we patch fields from the constructor, patch the getter/setter methods too.

Dexbolt
https://fburl.com/dexbolt/kfm09igr

Reviewed By: thezhangwei

Differential Revision: D57286831

fbshipit-source-id: d68bc7972c2252f9376064a2703d7f76f758af77