Patch lambda methods
Summary:
Lambda methods create a synthetic class. Values that are passed into the methods will become fields of the class, but they will not be annotated. If a parameter of the lambda is also an annotated parameter of the original method, patch the field of the synthetic class.

Some lambdas are nested heavily, so patch all the fields of the nested lambdas. By saving the common class name and the annotated field, we can patch all the other nested lambdas by searching up the common class name in the map.

Dexbolt
https://fburl.com/dexbolt/lrh6p9qy

Reviewed By: thezhangwei

Differential Revision: D57286832

fbshipit-source-id: a6cae9a6045f895d368e7e12e25b43c7ff83c648