Allow for creating interfaces with s-expr
Summary:
Add support for defining interfaces via `assembler::class_from_string()` with a similar form to regular classes.

Interfaces can be public or non-public, so access flags are still part of the required patterns, but access flags are omitted for methods/fields in the patterns (everything is public, methods are abstract and all fields are final/static). Improved pre-existing todos to allow s-expr classes to implement interfaces, and similarly for interfaces to extend one or many interfaces.

Not implemented here are encoded field values, that is left for a follow-up task.

Reviewed By: beicy

Differential Revision: D57295935

fbshipit-source-id: 741da24af2c7ba50eb9cac5b3ac18f55fc790357