Update interactions for deduped source blocks
Summary: If two duplicate blocks have the same number of source blocks, all the canonical block's source blocks' interactions should be the max of itself and the duplicate block. If they don't have the same number of source blocks, just take the max of the first and last ones for now.

Reviewed By: agampe

Differential Revision: D57463224

fbshipit-source-id: 02201d5180b75edbde2fa46e26af36ce9fa4d0c4