Add ReferencedState flag to prevent outlining code from a method
Summary:
Code that is being analyzed by simple, intraprocedural analysis might have problems running late in our pass list if outlining takes effect. The example of this that I have in mind is the autogenerated R classes, which we analyze <clinit> methods.

In some uncommon situations, outlined methods get created for very small instruction sequences, like creating an empty array of a constant size. Supporting method calls in a simple analysis, like OptimizeResourcesPass would be a large change, so making a simple way to configure no outlining for a given method would be a way to keep logic simple. The outlining in the examples I looked at seemed like a wash anyways and not providing much of a size benefit.

Reviewed By: thezhangwei

Differential Revision: D55669886

fbshipit-source-id: ccf6e486507adb4171c4f3733dd70defc52fe211