Bring back hard-coded exclusion for dalvik.annotation.MemberClasses
Summary: This annotation doesn't have any fields, so ignoring it doesn't cause any dangling references to removed classes, so it conceptually fine to ignore as we used to until I recently dropped this exclusion out of ignorance in D52986642. My change is fine as long as SystemAnnoKill runs before RMU, but if SystemAnnoKill is disabled, then it's quite beneficial for RMU to keep ignoring this annotation after all.

Reviewed By: ssj933

Differential Revision: D58302286

fbshipit-source-id: a5f99343f0fa7585ce9509a13ac22849d7127690