Add local variables in lambda checks
Summary:
Methods with the name `invoke` or `onClick` in synthetic classes are probably lambdas that might call a method that expects typedef annotated parameters. These methods need to propagate the annotation to the synthetic class' field and the class' constructor's parameters that IPUT to the field.

`patch_local_var_lambda` will check all the invoked methods within the `invoke` or `onClick` method. If the method is a default or access method, check the method that they call. This should be better than adding a dependency on having the default/access methods patched and needing to do another round of `walk::parallel::methods`.

After we finish patching the fields, we will walk through all the synthetic classes and patch their constructors in `patch_synth_cls_fields_from_ctor_param`. Some classes have multiple `invoke` or `onClick` methods. If we patch during `patch_local_var_lambda`, we will end up adding annotations to a parameter that's already annotated, which causes a failure.

Dexbolt:
https://fburl.com/dexbolt/evf2ixon

Reviewed By: thezhangwei

Differential Revision: D58224413

fbshipit-source-id: 14a2f4214bce8a420ddd90a049f4fca81bd68547