Add new keep reason for META-INF usages
Summary:
In some configs for our build targets, `record_keep_reasons` is set to true which kicks in extra logic when marking things as root.

Flip this on via `"record_keep_reasons": true` in the test cases, and fix an assert by making a new category for class names discovered under the META-INF/services folder in an app.

Reviewed By: ssj933

Differential Revision: D58536027

fbshipit-source-id: dd6f03d80c84998ee7340cb2efd364f9585f6efd