add option avoid_type_lookup_table_collisions
Summary:
The option is off by default, making this a behavior-preserving change.

If enabled, we model Android ART's data structure in `libdexfile/dex/type_lookup_table.cc` that is precomputed and used at runtime to lookup classes, in order to predict and avoid hash table collisions in those tables. We do this separately for each dex files, as each dex gets it's own lookup table in ART. When iteratively choosing the next class name in the RenameClassesV2 pass, we skip those (of the form `LX/abc;`) that would cause collisions within that dex.  However, as it turns out that we have to skip "most" such names (★), we do try to reuse previously skipped names in later dexes. So each dex still has monotonically increasing renamed class names, which is important for ordering of classes within each dex, but it is no longer the case that class names increase monotonically across all dexes. However, for non-root stores we still use the old scheme, as VoltronModuleMetadataHelper has certain assumptions about the consecutiveness of encoded indices in renamed class names.

(★) The hash function is the standard Java hash function that iteratively performs $h_{i+1} := h_i*31 + s[i]$ where $h_0 = 1$ and $h_n$ is the final hash for a string $s$ of length $n$. Given that we'd prefer to keep the size of our renamed classes short, i.e. 3 "digits" that change, this effectively only allows us to do a base-31 encoding for perfect hashing. Thus "most" encoded indices in the base-62 encoding that we traditionally use and that is expected in various places have to be skipped.

Reviewed By: thezhangwei

Differential Revision: D58160229

fbshipit-source-id: a6965d5b5412a70f588a004a47e20ea0fa01f1a3