Ensure no R class has code outlined
Summary:
Use the rstate flag added by D55669886 to keep the job that D53505200/D55265832 has to do much more simple.

Justification: there is hardly any value to outlining from the R class <clinit> methods; this was in practice happening not at all for some apps, or for other apps in a way that didn't seem to make the code much smaller anyways (outlined methods were simply creating a new array of a hard coded size, and calling such an outlined method looked like the same number of code units).

Reviewed By: ssj933

Differential Revision: D55670009

fbshipit-source-id: 40a4964b2c75c78cda87022e74847321eb702d44