Make R class identification much more specific
Summary:
Tighten our definition of what constitutes an R class. Goal is to make it possible to define a Scope of R classes that can participate in more sophisticated dce.

Given that R classes are expected to be very very simple autogenerated classes, with at most a <clinit> to set up its fields and empty constructor, put in a better effort in identifying such a class.

Reviewed By: ssj933

Differential Revision: D55265861

fbshipit-source-id: 71fb6e5f93bfc020a29257ab2e1d06a7373d2df5