Fix double remapping bug when array fields reference other fields
Summary:
There was a difficult to spot bug in this stack, where a R class that populates its int[] field via sget from some other R class was getting it's values remapped on top of an already remapped array. That's wrong, and caused this stack to cease to delete a couple of dead things.

The code in D53505200 was trying to do this all in 1 iteration, which lead to the mistake. Break it out into two loops, first to iterate in order gathering the base state, then remap individually.

Reviewed By: ssj933

Differential Revision: D59028256

fbshipit-source-id: 3dad9233ddb78037f732958a8c3799ed5a9fe369