option to derive stats
Summary:
On closer inspection of the pass, it would always make it so that all methods split out of hot methods would be regarded as cold as far as deep data and baseline profiles are concerned. That is because the pass disables itself during an "instrumentation" build, and thus all methods it generates would just not get tracked by deep data, and thus will not appear as ever called, and later get excluded in the baseline profiles in release builds.

This is getting addressed here by an adding an option to derive method-profile stats. The option is off for now, to make this diff a behavior-preserving change for easier testing.

See D57943267 for potential impact assessment.

Reviewed By: beicy

Differential Revision: D57943268

fbshipit-source-id: 7281aa7800d8303008699e409100aecaaf4fee5c