options to attach @NeverInline annotation
Summary:
dex2oat compiles hot code, as specified in the baseline profile, and while doing so inlines callees (with a certain bucket) -- regardless of whether the callees are hot or not. This can increase ARM64 code size by inlining cold callees into hot callers.

This diff introduces an option to annotate those cold callees which are called from hot methods with the NeverInline annotation which dex2oat respects.

As the options are off by default, this is a behavior-preserving change.

Differential Revision: D56367624

fbshipit-source-id: d3da3f38665191cb34818ec42e20df959a9fde9b