Patch annotated kotlin property getter/setters
Summary:
If a Kotlin property is annotated like the one below, we make sure that the synthesized get/setter is also patched with the matching Typedef annotation.
```
class Mismatch : BaseError() {
    TestStringDef override val failure: String = TestStringDef.ONE
  }
```

Doing so simplifies the case we ran into when checking a true virtual callsite. If the root of the virtual scope is an interface in Kotlin, with the upcoming change in D59025041, it can only be annotated on the getter, i.e., `get:MyTypeDefAnno`. The reason is that the property does not exist on the interface class. Whereas, all the subclass overrides can still have the annotation applied to the property directly, like `MyTypeDefAnno`. In this case, the annotation is only applied to the property field on the subclass, not on the getter/setters. However, the virtual call checking logic only looks at all the overriding methods. To make it work, we patch all property getter/setters if the matching property field is annotated with a Typedef annotation.

Reviewed By: itang00

Differential Revision: D59253644

fbshipit-source-id: 3062a1dc8c1240182f3dcc263ad0241830a9550f