Fix nondeterminism in typedefannochecker
Summary: There was nondeterminism in the previous implementation because we wouldn't always add param annotations to the same constructors if there were several first level nested lambdas with the same name. Now, run the entire method on all first level nested lambdas. This will cause over-patching, which will be resolved in the next diff

Reviewed By: thezhangwei

Differential Revision: D59307671

fbshipit-source-id: dfe6bef4ac650a362524aee1cb6a361efa0515c6