Patch annotated Kotlin Companion property accessor
Summary:
## Context

For a companion property written as below, the property `ONE` is actually placed on the `Listener` class itself. The access from the `getOne()` method on the companion object is enabled by a synthesized getter `access$getONE$cp()` on the outer `Listener` class.
```

  class Listener {
    companion object {
      TestStringDef private val ONE: String = TestStringDef.ONE

      TestStringDef
      fun getOne(): String {
        return ONE
      }
    }
  }
```
While the property `ONE` has the `TestStringDef` annotated attached as expected, the synthesized getter has no annotation. In this case, we want to patch it, so its caller can pass the safety checker.

## Solution

Identity such synthesized property accessor and patch it if the matching property field is annotated with Typedefs.

Reviewed By: itang00

Differential Revision: D59306992

fbshipit-source-id: 5bc14d1e24a746fff5c6ec9ed1c944fca14d5c1e