Update EnclosingClass annotation types for singly implemented interfaces
Summary:
Address another potential problem when running this pass without SystemAnnoKillPass. Inner types of the optimized away interface can have a dangling reference to the outer interface.

It's unclear what the best approach is under such a condition. Drop annotations? Update them to implementor? This change does the latter since it is similar to what was happening already.

Reviewed By: beicy

Differential Revision: D59294938

fbshipit-source-id: e534326a15d11dd25511a7bbaedb3f4ece054dcb