Copy over select annotations on interface during SingleImpl
Summary:
It seems that asking a class for its simple name will check some of the dalvik system annotations for enclosing classes.
- https://cs.android.com/android/platform/superproject/+/android-12.0.0_r1:libcore/ojluni/src/main/java/java/lang/Class.java;l=1253
- https://cs.android.com/android/platform/superproject/+/android-12.0.0_r1:libcore/ojluni/src/main/java/java/lang/Class.java;l=1357
- https://cs.android.com/android/platform/superproject/+/android-12.0.0_r1:art/runtime/native/java_lang_Class.cc;l=878
- https://cs.android.com/android/platform/superproject/+/android-12.0.0_r1:art/runtime/dex/dex_file_annotations.cc;l=1489

What SingleImpl does is kind of okay, but kind of misleading; it hauls over annotations on the interface class to the class which implements it (unless the implementor already had an annotation of that type).

Sensible at face value, but if you are the only class implementing [an interface which itself is an inner class](https://www.internalfb.com/code/fbsource/[b0b8ff3536a722b7b731d8e44024acccdf5aa27b]/fbandroid/libraries/onecamera/lib/src/main/java/com/facebook/onecamera/components/mediapipeline/gl/context/GlSurfaceFactory.kt?lines=21), it seems strange to copy over `EnclosingClass` / `InnerClass` annotations of the interface to the implementor, especially when the implementor was never an inner class to begin with.

The problem at hand, I believe is that these enclosing class annotations are getting left with a dangling type reference to a deleted type, which causes issues at runtime.

Reviewed By: beicy

Differential Revision: D59250713

fbshipit-source-id: a70da5aa0823ddcee5bc4f5defa708986db8c615