Patch annotated Kotlin private property getter
Summary:
In Kotlin, If a private property is accessed by an inner class, like a lambda class, a synthetic accessor/getter is produced to make it work. It seems like the patcher in the `TypedefAnnoChecker` only patches the synth getter if another synthesized `annotations` method exists.

Here we patch the private getter if the matching property is annotated with Typedef annotation.

Reviewed By: itang00

Differential Revision: D59550509

fbshipit-source-id: 969a1c6ea0bdfca22e1d4327fe7da134c827aeda