Improve documentation, method names for handling resource string data
Summary:
Many of these helper methods for reading string data came to be back when the version of libresource and the associated AOSP classes were on the very old forked version, and had buggy APIs that would return MUTF-8 data with the wrong length (it would return lengths as if it were UTF-16 which is misleading).

This lead to the approach where regardless of string pool encoding, the easiest/most reliable way to get data out was to force it to decode to UTF-16 and then convert to UTF-8. Easy to use, but error prone because this converts Android's Modified UTF-8 encoding to standard UTF-8 encoding! This means `arsc::get_string_from_pool()` can easily create a bug depending on the how it is used. D59091664 shows this.

The underlying deficiency in our import of AOSP code was addressed long ago in T111911654, but nevertheless, the helper method to do this remains in use. This change is to clarify comments, rename methods to make things more clear.

Overall, the conventions I want to put forward are:
- When modifying a string pool using `ResStringPoolBuilder` APIs, always copy along data from the original pool via the `add_string(const android::ResStringPool&, size_t)` overload: https://fburl.com/code/fbefb0rh
- For situations where string data is known to be only simple characters, like class names, attribute names, file paths in the zip, `arsc::get_string_from_pool()` is okay to use.
- For a consistent programming model between .apk and .aab inputs, returning standard UTF-8 string data from RedexResources.h APIs would be appropriate (thus, a pass would never have to care about the encoding and internal details are pushed down to ApkResources to get it right).
- For human readable situations (logging, dumping out to symbol files), it is okay to use `arsc::get_string_from_pool()`.

Reviewed By: nithyaarunmeta

Differential Revision: D59589674

fbshipit-source-id: 2e40a9f557f7f5c86883407ecd89d2feb8e9e6bf